Minor refactor, add handle to docs
